v13-evo: linking errors fixed, refactoring
