+++ src/clientversion.h
@@ -15,9 +15,9 @@
 
 //! These need to be macros, as clientversion.cpp's and dash*-res.rc's voodoo requires it
 #define CLIENT_VERSION_MAJOR 0
-#define CLIENT_VERSION_MINOR 12
+#define CLIENT_VERSION_MINOR 13
 #define CLIENT_VERSION_REVISION 0
-#define CLIENT_VERSION_BUILD 55
+#define CLIENT_VERSION_BUILD 0
 
 //! Set to true for release, false for prerelease or test build
 #define CLIENT_VERSION_IS_RELEASE true
