+++ src/evo/dapi.h
@@ -30,40 +30,6 @@
 using namespace std;
 using namespace json_spirit;
 
-std::string GetProfileFile(std::string strUID)
-{
-    boost::filesystem::path filename = GetDataDirectory() / "users" / strUID;
-    return filename.c_str();
-}
-
-std::string GetPrivateDataFile(std::string strUID, int nSlot)
-{
-    std::string strFilename = strUID + "." + boost::lexical_cast<std::string>(nSlot);
-    boost::filesystem::path filename = GetDataDirectory() / "users" / strFilename;
-    return filename.c_str();
-}
-
-
-std::string escapeJsonString(const std::string& input) {
-    std::ostringstream ss;
-    //for (auto iter = input.cbegin(); iter != input.cend(); iter++) {
-    //C++98/03:
-    for (std::string::const_iterator iter = input.begin(); iter != input.end(); iter++) {
-        switch (*iter) {
-            case '\\': ss << "\\\\"; break;
-            case '"': ss << "\\\""; break;
-            case '/': ss << "\\/"; break;
-            case '\b': ss << "\\b"; break;
-            case '\f': ss << "\\f"; break;
-            case '\n': ss << "\\n"; break;
-            case '\r': ss << "\\r"; break;
-            case '\t': ss << "\\t"; break;
-            default: ss << *iter; break;
-        }
-    }
-    return ss.str();
-}
-
 class CDAPI
 {
 private:
