+++ src/util.h
@@ -111,6 +111,13 @@ void AllocateFileRange(FILE *file, unsigned int offset, unsigned int length);
 bool RenameOver(boost::filesystem::path src, boost::filesystem::path dest);
 bool TryCreateDirectory(const boost::filesystem::path& p);
 boost::filesystem::path GetDefaultDataDir();
+boost::filesystem::path GetDataDirectory();
+
+std::string GetProfileFile(std::string strUID);
+std::string GetPrivateDataFile(std::string strUID, int nSlot);
+std::string escapeJsonString(const std::string& input);
+
+
 const boost::filesystem::path &GetDataDir(bool fNetSpecific = true);
 boost::filesystem::path GetConfigFile();
 boost::filesystem::path GetMasternodeConfigFile();
